+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.25
 
 * moved `distutils` to [snakehouse](https://pypi.org/project/snakehouse/)
+* added an option to write to logs in `DumpToFileHandler`